+++ tests/commands-anx-options/move-options.test.ts
@@ -14,6 +14,7 @@ describe('MoveOptions', () => {
     [{ '--branch': 'ref' }, ['--branch=ref']],
     [{ '--failed': null }, ['--failed']],
     [{ '--from': 'A' }, ['--from=A']],
+    [{ '--from-anywhere': null }, ['--from-anywhere']],
     [{ '--jobs': 2 }, ['--jobs=2']],
     [{ '--jobs': '3' }, ['--jobs=3']],
     [{ '--json': null }, ['--json']],