+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## vNext
+* Support git-annex 10.20231227 features
+    * Add CopyOptions --from-anywhere
+    * Add MoveOptions --from-anywhere
+
 ## v3.0.0 released 2023-10-11
 * Convert the package to an ECMAScript module from CommonJS
 * Add GitAnnexAPI methods